pyproject.toml
[project]
name = "hot100-pkg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "dash>=3.1.1",
    "dash-ag-grid>=31.3.1",
    "dash-core-components>=2.0.0",
    "dash-mantine-components>=2.1.0",
    "datetime>=5.5",
    "duckdb>=1.3.1",
    "httpx>=0.28.1",
    "httpx-retries>=0.4.0",
    "lxml>=5.4.0",
    "pathlib>=1.0.1",
    "polars[pyarrow]>=1.30.0",
    "pyyaml>=6.0.2",
    "tool>=0.8.0",
    "typing>=3.10.0.0",
]

[build-system]
requires = ["uv_build>=0.7.21,<0.8.0"]
build-backend = "uv_build"

[dependency-groups]
dev = []
